+++ R/utils.R
@@ -18,3 +18,16 @@ find_resource <- function(template, file) {
 ptrr_file <- function(...) {
   system.file(..., package = "ptrr", mustWork = TRUE)
 }
+
+copy_word_template <- function(template = "ptrr_word", destination = NULL) {
+  dest <- ifelse(is.null(destination), getwd(), file.path(getwd(), destination))
+  res <- find_resource(template, "template.docx")
+
+  file.copy(res, dest)
+
+  rslt <- ifelse(utils::file_test("-f", dest),
+                 file.path(dest),
+                 file.path(dest, basename(res))
+  )
+  return(rslt)
+}