+++ NEWS.md
@@ -1,5 +1,11 @@
 # ptrr (development version)
 
+* new basic Rmarkdown template
+* new date formatting functions
+* new ggplot2 utilities: themes, axes
+* new number formatters
+* new font utilities
+
 # ptrr 0.1.0
 
 * completed `insert_package_into_drat()`